dev/init_requirements.txt
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
smmap==5.0.1
